#CKFK20230409 Correcciones fecha

+++ app/src/main/java/com/dts/roadp/clsFinDia.java
@@ -88,9 +88,9 @@ public class clsFinDia extends PBase{
     }
 
     //#HS_20181121_1642 Esta función anteriormente pertenecía a la clase FinDia.
-    public int ultimoCierreFecha() {
+    public long ultimoCierreFecha() {
         Cursor DT;
-        int rslt=0;
+        long rslt=0;
 
         try {
             sql="SELECT val1 FROM FinDia";
@@ -98,7 +98,7 @@ public class clsFinDia extends PBase{
 
             if(DT.getCount()>0){
                 DT.moveToFirst();
-                rslt=DT.getInt(0);
+                rslt=DT.getLong(0);
             }
         } catch (Exception e) {
             addlog(new Object(){}.getClass().getEnclosingMethod().getName(),e.getMessage(),sql);
